refactor: moce to viralunity folder
